move samplerum to utils.js
